Refactor map to its own class
