+++ frontend/src/views/Search.vue
@@ -164,14 +164,6 @@
     &:focus
       outline: thin dotted
 
-  .map
-    height: 300px
-    width: 300px
-    margin-bottom: $filter-margin
-
-  .wrapper
-    position: relative
-
   .no-padding
     padding: 0
 
@@ -253,13 +245,11 @@
   <section id="sideBar">
     <header class="filterOptions">Filter search</header>
 
-    <div id="map" ref="mapElement" class="container wrapper" style="z-index: 4">
-      <div class="row">
-        <div class="col-md-12 no-padding">
-          <div id="map" class="map"></div>
-        </div>
-      </div>
-    </div>
+    <Map v-if="allSites && allSites.length > 0"
+      :sites="allSites"
+      :selectedSiteIds="selectedSiteIds"
+      :onMapMarkerClick="onMapMarkerClick">
+    </Map>
 
     <custom-multiselect
       label="Location"
@@ -415,15 +405,15 @@ import { BPagination } from 'bootstrap-vue/esm/components/pagination'
 import Datepicker from '../components/Datepicker.vue'
 import CustomMultiselect from '../components/Multiselect.vue'
 import DataSearchResult from '../components/DataSearchResult.vue'
-import { dateToString, getIconUrl, getShadowUrl, getMarkerUrl, humanReadableSize, combinedFileSize,
+import { dateToString, getIconUrl, humanReadableSize, combinedFileSize,
   fixedRanges, getDateFromBeginningOfYear, isSameDay} from '../lib'
 import { DevMode } from '../lib/DevMode'
 import VizSearchResult from '../components/VizSearchResult.vue'
 import {Visualization} from '../../../backend/src/entity/Visualization'
 import {Product} from '../../../backend/src/entity/Product'
 import {ProductVariable} from '../../../backend/src/entity/ProductVariable'
-import L, { marker } from 'leaflet'
 import {SearchFileResponse} from '../../../backend/src/entity/SearchFileResponse'
+import Map from '@/components/Map.vue'
 
 Vue.component('datepicker', Datepicker)
 Vue.component('b-table', BTable)
@@ -439,7 +429,9 @@ export interface Selection {
   humanReadableName: string;
 }
 
-@Component({name: 'app-search'})
+@Component({name: 'app-search',
+  components: {Map}
+})
 export default class Search extends Vue {
   @Prop() mode!: string
 
@@ -500,33 +492,6 @@ export default class Search extends Vue {
       .flatMap(prod => prod.variables)
   }
 
-  // map
-  map: L.Map | null = null
-  tileLayer: L.TileLayer | null = null
-  allMarkers: { [key: string]: L.Marker } = {}
-  passiveMarker = L.Icon.extend({
-    options: {
-      iconUrl: getMarkerUrl('blue'),
-      iconSize: [25, 41],
-      iconAnchor: [12, 41],
-      popupAnchor: [1, -34],
-      shadowUrl: getShadowUrl(),
-      shadowSize: [41, 41],
-      shadowAnchor: [12, 41]
-    }
-  })
-  activeMarker = L.Icon.extend({
-    options: {
-      iconUrl: getMarkerUrl('green'),
-      iconSize: [25, 41],
-      iconAnchor: [12, 41],
-      popupAnchor: [1, -34],
-      shadowUrl: getShadowUrl(),
-      shadowSize: [41, 41],
-      shadowAnchor: [12, 41]
-    }
-  })
-
   renderComplete = false
 
   displayBetaNotification = true
@@ -577,31 +542,7 @@ export default class Search extends Vue {
     this.$nextTick(() => {
       this.renderComplete = true
     })
-    this.initMap()
     this.addKeyPressListener()
-  }
-
-  initMap() {
-    this.map = L.map(this.$refs['mapElement'] as HTMLElement).setView([54.00, 14.00], 3)
-    this.tileLayer = L.tileLayer('https://cartodb-basemaps-{s}.global.ssl.fastly.net/rastertiles/voyager/{z}/{x}/{y}.png')
-    this.tileLayer.addTo(this.map)
-  }
-
-  initLayers()  {
-    const markerNames = this.allSites.map(site => site.humanReadableName)
-    const lat = this.allSites.map(site => site.latitude)
-    const lon = this.allSites.map(site => site.longitude)
-    const id = this.allSites.map(site => site.id)
-    markerNames.forEach((name, i) => {
-      const mark = marker([lat[i], lon[i]])
-      mark.setIcon(new this.passiveMarker)
-      mark.on('click', (_onClick) => {
-        this.onMapMarkerClick(id[i])
-      })
-      this.allMarkers[id[i]] = mark
-      if (!this.map) return
-      mark.addTo(this.map)
-    })
   }
 
   onMapMarkerClick(id: string) {
@@ -611,19 +552,6 @@ export default class Search extends Vue {
     else {
       this.selectedSiteIds.push(id)
     }
-  }
-
-  setActiveMarkers() {
-    const keys = Object.keys(this.allMarkers)
-    keys.forEach((id: string) => {
-      const mark = this.allMarkers[id]
-      if (this.selectedSiteIds.includes(id)) {
-        mark.setIcon(new this.activeMarker)
-      }
-      else {
-        mark.setIcon(new this.passiveMarker)
-      }
-    })
   }
 
   alphabeticalSort = (a: Selection, b: Selection) => a.humanReadableName > b.humanReadableName
@@ -636,7 +564,6 @@ export default class Search extends Vue {
     ]).then(([sites, products]) => {
       this.allSites = sites.data.sort(this.alphabeticalSort)
       this.allProducts = products.data.sort(this.alphabeticalSort)
-      this.initLayers()
       if (this.isVizMode()) {
         this.selectedSiteIds.push('bucharest')
         this.selectedProductIds.push('classification')
@@ -777,7 +704,6 @@ export default class Search extends Vue {
   @Watch('selectedSiteIds')
   onSiteSelected() {
     this.fetchData()
-    this.setActiveMarkers()
   }
 
   @Watch('dateFrom')
